[Decomp][Spell] Cohesion split: extract SetTargetMap to SpellTargeting.cpp
Move Spell::SetTargetMap and its file-local sort helpers (ChainHealingOrder,
ChainHealingFullHealth, TargetDistanceOrderNear, binary_function,
TargetDistanceOrderFarAway) out of Spell.cpp into a new SpellTargeting.cpp.
The helpers are used only by SetTargetMap, so they travel with it.

Also moves the FindCorpseUsing<T> template definition into the same TU:
SetTargetMap is its only instantiation site tree-wide, and the template body
lives in a .cpp (not a header), so the body must share the TU with its
instantiator. Declaration stays in Spell.h.

Bodies only; all declarations stay in Spell.h, so call sites and linkage are
unchanged. Spell.cpp 10806 -> 9048. Pure-deletion diff on Spell.cpp.
